#123 refactor: DTO 클래스 리팩토링 및 패키지 구조 변경
DTO 클래스 리팩토링 및 패키지 구조 변경
- TripResponseDTO를 MapViewResponseDTO로 이름 변경
- PointImageDTO를 PinPointImageGalleryResponseDTO로 이름 변경
- TripInfoRequestDTO, TripInfoResponseDTO의 위치를 model에서 dto 패키지로 이동
- 불필요한 메서드 삭제 및 관련 코드 수정
