+++ src/main/java/com/fivefeeling/memory/domain/trip/dto/MapViewResponseDTO.java
@@ -1,0 +1,15 @@
+package com.fivefeeling.memory.domain.trip.dto;
+
+import com.fivefeeling.memory.domain.media.model.MediaFileResponseDTO;
+import com.fivefeeling.memory.domain.pinpoint.model.PinPointResponseDTO;
+import java.util.List;
+
+public record MapViewResponseDTO(
+        String tripTitle,
+        String startDate,
+        String endDate,
+        List<PinPointResponseDTO> pinPoints,
+        List<MediaFileResponseDTO> mediaFiles
+) {
+
+}
