+++ src/main/java/com/fivefeeling/memory/domain/trip/dto/TripsResponseDTO.java
@@ -2,6 +2,6 @@ package com.fivefeeling.memory.domain.trip.dto;
 
 import java.util.List;
 
-public record TripsResponseDTO(List<TripInfoResponseDTO> trips) {
+public record TripsResponseDTO(List<TripResponseDTO> trips) {
 
 }
