--- src/main/java/com/fivefeeling/memory/domain/trip/model/TripInfoRequestDTO.java
@@ -1,16 +1,0 @@
-package com.fivefeeling.memory.domain.trip.model;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import java.time.LocalDate;
-import java.util.List;
-
-public record TripInfoRequestDTO(
-    @NotBlank String tripTitle,
-    @NotBlank String country,
-    @NotNull LocalDate startDate,
-    @NotNull LocalDate endDate,
-    List<String> hashtags
-) {
-
-}
