--- src/main/java/com/fivefeeling/memory/domain/trip/model/TripSummaryDTO.java
@@ -1,9 +1,0 @@
-package com.fivefeeling.memory.domain.trip.model;
-
-public record TripSummaryDTO(
-        int tripCount,
-        TripInfoResponseDTO recentlyTrip
-) {
-
-}
-
